+++ napalm_yang/helpers.py
@@ -1,11 +1,4 @@
 import yaml
-
-from napalm_yang.parsers.text import TextParser
-from napalm_yang.parsers.xml import XMLParser
-
-from napalm_yang.translators.text import TextTranslator
-from napalm_yang.translators.xml import XMLTranslator
-
 import os
 import jinja2
 
@@ -24,16 +17,6 @@ def yaml_include(loader, node):
 
 
 yaml.add_constructor("!include", yaml_include)
-
-
-def get_parser(parser):
-    parsers = {
-        "TextParser": TextParser,
-        "XMLParser": XMLParser,
-        "TextTranslator": TextTranslator,
-        "XMLTranslator": XMLTranslator,
-    }
-    return parsers[parser]
 
 
 def find_yang_file(profile, filename, path):
